+++ go.mod
@@ -6,16 +6,12 @@ require (
 	github.com/rs/zerolog v1.34.0
 	github.com/stretchr/testify v1.9.0
 	go.dedis.ch/kyber/v4 v4.0.0-pre2.0.20250219110603-23debab3f61d
-	go.dedis.ch/protobuf v1.0.11
 	golang.org/x/xerrors v0.0.0-20191204190536-9bdfabe68543
-	google.golang.org/protobuf v1.36.6
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/kr/pretty v0.3.1 // indirect
-	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/mattn/go-isatty v0.0.19 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.19 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
