+++ networking/fake_network.go
@@ -74,19 +74,6 @@ func (n *FakeNetwork) Broadcast(msg []byte, from int64) error {
 	return nil
 }
 
-// NetworkInterface represents an interface used by a node to communicate in the network
-type NetworkInterface interface {
-	// Send allows to send a byte message to a recipient addressed by an int
-	Send([]byte, int64) error
-	// Broadcast send the given byte message to everyone else in the network
-	Broadcast([]byte) error
-	// Receive waits on the channel for a message to arrive. Blocks until a message arrives or
-	// the channel is written to. This allows stopping before receiving a message
-	Receive(context.Context) ([]byte, error)
-	GetID() int64
-	GetSent() [][]byte
-	GetReceived() [][]byte
-}
 type FakeInterface struct {
 	rcvQueue     chan []byte
 	sendMsg      func([]byte, int64, int64) error
